:lipstick: Force a full refresh of the plot when the window gets a resize
It looks like there might be something odd either about how textual-plotext
is working, or how sizes and calls to render work out when a resize happens,
but this seems necessary here for now.

When I remember I'll dive into this more and raise an issue in the
textual-plotext repo to remind me.